+++ Source/Bugsnag.m
@@ -56,6 +56,9 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
 @interface BugsnagClient ()
 - (void)startListeningForStateChangeNotification:(NSString *_Nonnull)notificationName;
 - (void)addBreadcrumbWithBlock:(void (^_Nonnull)(BugsnagBreadcrumb *_Nonnull))block;
+- (void)internalClientNotify:(NSException *_Nonnull)exception
+                    withData:(NSDictionary *_Nullable)metadata
+                       block:(BugsnagOnErrorBlock _Nullable)block;
 @end
 
 @interface BugsnagMetadata ()
@@ -148,6 +151,10 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
     }
 }
 
+/**
+ * Intended for use by other clients (React Native/Unity). Calling this method
+ * directly from iOS is not supported.
+ */
 + (void)internalClientNotify:(NSException *_Nonnull)exception
                     withData:(NSDictionary *_Nullable)metadata
                        block:(BugsnagOnErrorBlock _Nullable)block {
@@ -301,6 +308,13 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
     if ([self bugsnagStarted]) {
         [self.client setContext:context];
     }
+}
+
++ (NSString *_Nullable)context {
+    if ([self bugsnagStarted]) {
+        return self.client.context;
+    }
+    return nil;
 }
 
 + (BugsnagUser *)user {