+++ Source/BugsnagConfiguration.h
@@ -330,12 +330,4 @@ typedef NS_OPTIONS(NSUInteger, BSGEnabledErrorType) {
 
 - (void)addPlugin:(id<BugsnagPlugin> _Nonnull)plugin;
 
-/**
- * Should the specified type of breadcrumb be recorded?
- *
- * @param type The type of breadcrumb
- *
- * @returns A boolean indicating whether the specified breadcrumb type should be recorded
- */
-- (BOOL)shouldRecordBreadcrumbType:(BSGBreadcrumbType)type;
 @end