+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/iOSTestApp-Bridging-Header.h
@@ -5,3 +5,4 @@
 #import <Bugsnag/Bugsnag.h>
 #import "Scenario.h"
 #import "AttachCustomStacktraceHook.h"
+#import "BugsnagHooks.h"