+++ CHANGELOG.md
@@ -11,8 +11,23 @@ Bugsnag Notifiers on other platforms.
 * Make `BugsnagOnErrorBlock` return `BOOL` rather than `void`
   [#555](https://github.com/bugsnag/bugsnag-cocoa/pull/555)
 
+* Make `BugsnagPlugin` take `BugsnagClient` as param
+  [#558](https://github.com/bugsnag/bugsnag-cocoa/pull/558)
+
+* Make user editable on `BugsnagEvent`
+  [#557](https://github.com/bugsnag/bugsnag-cocoa/pull/557)
+
+* Add getter for `Bugsnag.context`
+  [#554](https://github.com/bugsnag/bugsnag-cocoa/pull/554)
+
+* Make `BugsnagUser` properties readonly
+  [#556](https://github.com/bugsnag/bugsnag-cocoa/pull/556)
+
 * Add `sendThreads` property to `BugsnagConfiguration`
   [#549](https://github.com/bugsnag/bugsnag-cocoa/pull/549)
+
+* Hide additional methods from public API
+  [#552](https://github.com/bugsnag/bugsnag-cocoa/pull/552)
 
 * Add structured app/device fields to `BugsnagSession`
   [#546](https://github.com/bugsnag/bugsnag-cocoa/pull/546)